requirements.txt
pdfplumber>=0.10.0
PyMuPDF>=1.24.0
rapidocr-onnxruntime>=1.3.0
Pillow>=10.0.0
ebooklib>=0.18
python-docx>=1.0.0
beautifulsoup4>=4.12.0
chromadb>=0.4.0
anthropic>=0.25.0
zhipuai>=2.0.0
openai>=1.10.0
click>=8.1.0
langchain-text-splitters>=0.0.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
lxml>=5.0.0
yfinance>=0.2.0
finnhub-python>=2.4.0
flask>=3.0.0
flask-cors>=4.0.0
# Bloomberg API (需要本地 Terminal)
# pip install --index-url=https://bloomberg.bintray.com/pip/simple blpapi
